requirements.txt
opencv-python
Pillow
ultralytics
pyttsx3